backend/requirements.txt
fastapi
pymongo
PyMySQL
redis
requests
Scrapy
SQLAlchemy
scrapy-redis
uvicorn
PyJWT
passlib
python-jose
python-dotenv
motor